Access control #28
Access control #28

+++ sso-web-api/src/main/java/com/study/infosecurity/ssoByRoles/contollers/MotorcycleController.java
@@ -2,10 +2,13 @@ package com.study.infosecurity.ssoByRoles.contollers;
 
 
 import com.study.infosecurity.ssoByRoles.model.dto.Motorcycle;
+import com.study.infosecurity.ssoByRoles.model.dto.User;
 import com.study.infosecurity.ssoByRoles.model.poko.constant.ResponseCode;
+import com.study.infosecurity.ssoByRoles.model.poko.constant.Roles;
 import com.study.infosecurity.ssoByRoles.model.poko.response.CommonResponse;
 import com.study.infosecurity.ssoByRoles.model.poko.response.MotorcycleResponse;
 import com.study.infosecurity.ssoByRoles.service.MotorcycleService;
+import org.apache.catalina.servlet4preview.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,6 +22,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class MotorcycleController {
 
     private MotorcycleService motorcycleService;
+    private String USER = "user";
+
+    @Autowired
+    private HttpServletRequest request;
 
     @Autowired
     public MotorcycleController(MotorcycleService motorcycleService) {
@@ -29,17 +36,20 @@ public class MotorcycleController {
     public CommonResponse getAllMotorcycles() {
         try{
             return new MotorcycleResponse( motorcycleService.findAll());
-    }
-    catch (Exception ex){
-        return new CommonResponse(ResponseCode.ERROR, "Some Error =(");
-    }
+        }
+        catch (Exception ex){
+            return new CommonResponse(ResponseCode.ERROR, "Some Error =(");
+        }
     }
 
     @RequestMapping(value = "/addMotorcycle", method = RequestMethod.POST)
     public CommonResponse addMotorcycle(@RequestBody Motorcycle motorcycle) {
         try{
-            motorcycleService.save(motorcycle);
-            return new CommonResponse(ResponseCode.SUCCESS, "Motorcycle succesfully added");
+            if(this.hasAccess()){
+                motorcycleService.save(motorcycle);
+                return new CommonResponse(ResponseCode.SUCCESS, "Motorcycle succesfully added");
+            }
+            return new CommonResponse(ResponseCode.ERROR, "Access Denied");
         }
         catch (Exception ex){
             return new CommonResponse(ResponseCode.ERROR, "Some Error =(");
@@ -49,7 +59,7 @@ public class MotorcycleController {
     @RequestMapping(value = "/updateMotorcycle", method = RequestMethod.GET)
     public CommonResponse updateMotorcycle(@RequestBody Motorcycle motorcycle) {
         try{
-            if (motorcycleService.exists(motorcycle.getId())){
+            if (motorcycleService.exists(motorcycle.getId()) && hasAccess()){
                 motorcycleService.save(motorcycle);
                 return new CommonResponse(ResponseCode.SUCCESS, "Motorcycle succesfully udated");
             }
@@ -63,11 +73,23 @@ public class MotorcycleController {
     @RequestMapping(value = "/deleteMotorcycle", method = RequestMethod.GET)
     public CommonResponse deleteMotorcycle(@RequestParam String id) {
         try{
-            motorcycleService.deleteById(Long.parseLong(id));
-            return new CommonResponse(ResponseCode.SUCCESS, "Motorcycle succesfully deleted");
+            if(this.hasAccess()){
+                motorcycleService.deleteById(Long.parseLong(id));
+                return new CommonResponse(ResponseCode.SUCCESS, "Motorcycle succesfully deleted");
+            }
+            return new CommonResponse(ResponseCode.ERROR, "Access Denied");
         }
         catch (Exception ex){
             return new CommonResponse(ResponseCode.ERROR, "Some Error =(");
         }
+    }
+
+    private boolean hasAccess(){
+        User user = (User) this.request.getAttribute(USER);
+        Roles item = user.getRoles().stream()
+                .filter(e -> e.equals(Roles.MOTORCYCLE_WRITE))
+                .findFirst()
+                .orElse(null);
+        return item != null;
     }
 }
